feat(clients): scope Call Centre tab to the client's own group
Extract Sales Hub's call-center logic into a reusable CallCentreContent
component (same clientGroups/showGroupFilter/showHeader pattern as
MarketingContent/LeadsContent), then wire it into the per-client internal
page's "Call Centre" tab (previously a ComingSoon stub), scoped to that
client's single group. The Overview and Members tabs are hidden in scoped
mode since they aren't meaningful (or safe) to show for a single client.
